Adding the functionality of showing different colors based on the done value according to the target value.

+++ lib/repository/habit_repo.dart
@@ -92,7 +92,7 @@ class HabitRepo {
 
       if (checkDatePresent(currDate, habit)) {
         totalWeeklyProgress.value +=
-            int.parse(getDoneValueForMeasurableHabit(currDate, habit)!);
+            double.parse(getDoneValueForMeasurableHabit(currDate, habit)!);
       }
     }
